Experimentally use JSString backing store for DOMString.
If the Latin-1 representation of the string is ASCII, we can use it as the representation of a Rust String.

This is unsafe for a couple of reasons:

a) SM might GC or move the string.
b) Servo might mutate the string.

So this code is just for timing experiments, not production use!